Fix SDL2 port of EventDispatcher class
* EventDispatcher: Change ERR check to match SDL2 API
* Input: Argument 'type' is now uint32 (from uint8) as per SDL2 API

+++ include/nomlib/system/Input.hpp
@@ -91,7 +91,7 @@ class Input
 
     virtual void HandleInput ( SDL_Event* event );
   protected:
-    virtual void onUserEvent ( uint8 type, int32 code, void* data1, void* data2 );
+    virtual void onUserEvent ( uint32 type, int32 code, void* data1, void* data2 );
 
     virtual void onQuit ( void );
     virtual void onResize ( int32 width, int32 height );
